+++ src/main/java/com/example/Dblog/board/BoardCreateForm.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotEmpty;
 @Getter
 @Setter
 public class BoardCreateForm {
-    @NotEmpty(message = "사용자 이름은 필수항목입니다.")
+    @NotEmpty(message = "작성자는 필수항목입니다.")
     private String user;
 
     @NotEmpty(message = "글 제목은 필수항목입니다.")
