+++ src/main/java/com/example/Dblog/board/BoardService.java
@@ -2,10 +2,8 @@ package com.example.Dblog.board;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -14,23 +12,20 @@ public class BoardService {
     private final BoardRepository boardRepository;
 
     @Transactional
-    public boolean cretePost(BoardCreateForm form, Long fileId){
+    public void cretePost(BoardCreateForm form, Long fileId){
         BoardEntity board = new BoardEntity();
         board.setUser(form.getUser());
         board.setTitle(form.getTitle());
         board.setContent(form.getContent());
         board.setFileid(fileId);
         this.boardRepository.save(board);
-        return true;
     }
-
     @Transactional
-    public boolean cretePost(BoardCreateForm form){
+    public void cretePost(BoardCreateForm form){
         BoardEntity board = new BoardEntity();
         board.setUser(form.getUser());
         board.setTitle(form.getTitle());
         board.setContent(form.getContent());
         this.boardRepository.save(board);
-        return true;
     }
 }
